refactor home view and pull the server list out into it's own component
